chore: add chart.version field and unify the indentation style of the config files
