+++ docs/plugins/devlake.md
@@ -1,9 +1,9 @@
-# apache-devlake plugin
+# devlake plugin
 
 //TODO(daniel-hutao): add document here after Chinese version be ready.
 
 ## Usage
 
 ``` yaml
---8<-- "apache-devlake.yaml"
+--8<-- "devlake.yaml"
 ```
