+++ docs/plugins/devlake.zh.md
@@ -1,9 +1,9 @@
-# apache-devlake 插件
+# devlake 插件
 
 //TODO(daniel-hutao): 本插件出于开发中状态。
 
 ## 用例
 
 ``` yaml
---8<-- "apache-devlake.yaml"
+--8<-- "devlake.yaml"
 ```
