+++ internal/pkg/plugin/devlake/create.go
@@ -1,4 +1,4 @@
-package apachedevlake
+package devlake
 
 import (
 	"github.com/devstream-io/devstream/internal/pkg/plugininstaller"
@@ -6,14 +6,16 @@ import (
 	"github.com/devstream-io/devstream/pkg/util/log"
 )
 
-func Read(options map[string]interface{}) (map[string]interface{}, error) {
+func Create(options map[string]interface{}) (map[string]interface{}, error) {
 	// Initialize Operator with Operations
 	operator := &plugininstaller.Operator{
 		PreExecuteOperations: plugininstaller.PreExecuteOperations{
 			helm.SetDefaultConfig(&defaultHelmConfig),
 			helm.Validate,
 		},
-		GetStateOperation: helm.GetPluginAllState,
+		ExecuteOperations:   helm.DefaultCreateOperations,
+		TerminateOperations: helm.DefaultTerminateOperations,
+		GetStateOperation:   helm.GetPluginAllState,
 	}
 
 	// Execute all Operations in Operator
