+++ internal/pkg/plugin/devlake/delete.go
@@ -1,27 +1,25 @@
-package apachedevlake
+package devlake
 
 import (
 	"github.com/devstream-io/devstream/internal/pkg/plugininstaller"
 	"github.com/devstream-io/devstream/internal/pkg/plugininstaller/helm"
-	"github.com/devstream-io/devstream/pkg/util/log"
 )
 
-func Update(options map[string]interface{}) (map[string]interface{}, error) {
+func Delete(options map[string]interface{}) (bool, error) {
 	// Initialize Operator with Operations
 	operator := &plugininstaller.Operator{
 		PreExecuteOperations: plugininstaller.PreExecuteOperations{
 			helm.SetDefaultConfig(&defaultHelmConfig),
 			helm.Validate,
 		},
-		ExecuteOperations: helm.DefaultUpdateOperations,
-		GetStateOperation: helm.GetPluginAllState,
+		ExecuteOperations: helm.DefaultDeleteOperations,
 	}
 
 	// Execute all Operations in Operator
-	status, err := operator.Execute(plugininstaller.RawOptions(options))
+	_, err := operator.Execute(plugininstaller.RawOptions(options))
 	if err != nil {
-		return nil, err
+		return false, err
 	}
-	log.Debugf("Return map: %v", status)
-	return status, nil
+
+	return true, nil
 }
