+++ internal/pkg/show/config/plugins/argocd.yaml
@@ -19,6 +19,8 @@ tools:
         chartPath: ""
         # name of the chart
         chartName: argo/argo-cd
+        # version of the chart
+        version: ""
         # release name of the chart
         releaseName: argocd
         # k8s namespace where ArgoCD will be installed
